requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
schedule==1.2.1
selenium==4.16.0
webdriver-manager==4.0.1
ShopifyAPI==12.4.0
geckodriver-autoinstaller==0.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23 